pyproject.toml
[project]
name = "csg-format"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.4",
    "jwt>=1.3.1",
    "libsql-experimental>=0.0.41",
    "python-dotenv>=1.0.1",
]
